fix #1
There could be some bug with the 'import' mechanism in vim9script.

If `g:ScopePopupOptionsSet` is called in the vimrc, the file will be
correctly imported in autoload/scope/fuzzy.vim as well. If the function
isn't called, fuzzy.vim won't be able to import popup.vim. I think
because the import is tagged as 'autoload' the first time, and fuzzy.vim
assumes (wrongly) that the script has been sourced already.

For now, move the setup function in the popup script, so that it is'nt
imported. Moreover, using a function defined in the plugin/ script
forces to use `packadd` instead of `packadd!`, otherwise the global
function isn't available in the vimrc.